// src/app/shared/interfaces/register-form.ts
export interface RegisterForm {
  pPlano: number;
  pVoucher: string;
  pCupom: string;
  pNome: string;
  pEmail: string;
  pSexo: string;
  pCpf: string;
  pTelefone: string;
  pNascimento: string;
  pCep: string;
  pPais: string;
  pEstado: string;
  pCidade: string;
  pEndereço: string;
  pNumero: string;
  pComplemento: string;
  pSenha: string;
}